+++ library/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.5.0 (July 28, 2023)
 
 - Fix invalid `comparable` import when using Deno
 - Add util functions to exports of library